Began Renaming code and moved nonrelevant code to its own package in order to clean up this mess and make it easy to understand
