+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/RedPark.java
@@ -1,7 +1,8 @@
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.disabled;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -18,9 +19,9 @@ import org.openftc.easyopencv.OpenCvCameraRotation;
 import org.openftc.easyopencv.OpenCvInternalCamera;
 import org.openftc.easyopencv.OpenCvPipeline;
 
-
+@Disabled
 @Autonomous
-public class B2SpinAndPark extends LinearOpMode {
+public class RedPark extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -87,13 +88,8 @@ public class B2SpinAndPark extends LinearOpMode {
 
             encoderDriveWithTimeLeft(.5,.5);
             sleep(100);
-            encoderDriveWithTimeForward(7,.5);
+            encoderDriveWithTimeForward(3.7,1);
             sleep(100);
-            spinRight(7, .3);
-            sleep(100);
-            encoderDriveWithTimeLeft(.4,.5);
-            sleep(100);
-            encoderDriveWithTimeBackward(6,1);
             stop();
 
             double moveLength = 28;
