+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/coopteleop2.java
@@ -1,8 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.disabled;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
+
+import org.firstinspires.ftc.teamcode.zanehardware;
 
 @TeleOp(name = "Teleop Tank Drive Coop that would probaly work")
 public class coopteleop2 extends LinearOpMode {
